ListenerMode1/MessyButPractical/TAICA/requirements.txt
firebase-admin
fastapi
uvicorn
python-decouple
twilio
python-dotenv
requests
openai
python-multipart
pydub






